pyproject.toml
[project]
name = "fasthtml-uv-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["python-fasthtml", "monsterui"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
